Implement processor API similar to assembler
* Move processor out or `mod.rs`
* Return `MvnReportError` on processor function calls
* Propagate errors to the `process` function in `mod.rs`
